Renamed vimrc and gvimrc files
I renamed the vimrc and gvimrc files to add a .vim at the end, in an
attempt of fixing the problem that GitHub has with it's syntax
highlighting and not knowing what language it is.

I also updated the symlink maker in order to reflect the new source
files for the .vimrc and .gvimrc symlinks.