First commit on new Arch Linux! :D
I just did some changes to the Python symlink script so that it works
correctly.

I added a new xinitrc file, which is part of Linux, so it was added to
the linux folder.

I did some changes to the setup_archlinux.sh file.

And finally some changes to the vimrc, nothing major.